Address code review comments.
This isn't exactly what Jason suggested. The reason for that is that
I couldn't find a way to simply wrap the ParameterizedFilter.filter
method in the HaystackFilter.filter method. Each of the three filter
methods have three steps that must be done in this order: (1) determine
that this filter is active, (2) apply a parameter or haystack search
(skip this step for the StandardFilter) and (3) perform standard
filtering and sorting operations. Since step (2) is the one which has
variations, it made sense to separate out (1) and (3) as methods of
the base class.