Use HTTP headers to mitigate database save conflicts.
HTTP has a feature called ETags (see http://en.wikipedia.org/wiki/HTTP_ETag)
that we can use to verify that a HTTP PUT is modifying something that has not
yet been modified by somebody else. When we perform a GET of a resource, the
response would include the JSON we get now, and it would also have a HTTP
header setting the ETag to the hex digest of the SHA-1 hash of the JSON. Savory
Pie must provide this header in response to a GET request.

Savory Pie responds to a HTTP PUT inside a MySQL transaction. If the HTTP PUT
includes an If-Match header with the expected hex digest for that resource,
Savory Pie should get the resource as JSON and compute the SHA-1, and if there
is a mismatch, the PUT fails with a HTTP 412 (precondition failed, see
http://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html). If it matches, the PUT
proceeds normally, producing a HTTP 204. (Unless there is a validation failure
which results in a HTTP 400.)

The client will typically perform a GET on a resource while setting up a HTML
form page where the user can edit the fields of the resource, and the form page
will have a submit button. When the user hits the submit button, the client
should PUT with an If-Match header with the hash that it received from the GET.
The PUT will fail with a 412 status if the resource has been modified by
another user in the meanwhile. Based on this, the client can respond in some
intelligent way to the conflict.

Something that will work in many cases would be to save the data from the first
GET, so that the client knows what the user has edited, and then perform a
second GET to see what the other user changed. If the changes are
non-overlapping and can therefore be merged, PUT the merged changes using the
hash from the second GET in the If-Match header. If this second PUT attempt
fails, probably the best recourse is to recommend the user preserve their
changes outside the browser (e.g. copy and paste to a text editor), refresh the
page, and re-apply the changes.