Fix save-conflict-mitigation issues found in subsequent testing.
The first issue is that a PUT may include PUTs of sub-resources, and I was only
looking at a single hash where the sub-resources would have different hashes.
The correct fix is to check the hash on _process_put (the outermost PUT) rather
than ModelResource.put.  Moving the check to _process_put enabled some
simplifications of the code (elimination of a separate exception, and the
expected hash doesn't need to be an attribute of the APIContext).

The second issue is we sometimes do single-resource GETs, but we also do a
QuerySetResource query of a number of resources, and in that case we want to
get back hashes for all of them. So now the response to such a query looks like
this:

  {
    meta: {count: N, ...},
    objects: [{...resource 1..., $hash: '1234...'},
              {...resource 2..., $hash: '5678...'},
              ...et cetera...
              {...resource N..., $hash: 'abcd...'}],
  }

These hashes can be used in If-Match headers to PUT to these resources.

The Django request key for finding the 'If-Match' header is actually
'HTTP_IF_MATCH'.

Also, validation errors (HTTP 400) should take precedence over hash mismatch
errors (HTTP 412).