Database transaction contention -> HTTP 409
We sometimes see a bunch of requests all come in at about the same time.  This
is the case if a piece of front-end Javascript takes a list of objects to PUT
or POST, and launches parallel AJAX requests for all of them. The first to
arrive will lock the database because POSTs and PUTs are now wrapped in
database transactions, and the second finds the database locked and raises a
TransactionManagementError. The previous behavior was to return a HTTP 500,
but that's what we get from a lot of other possible server-side failures.

This situation is frequently recoverable so we want to return a HTTP status
code (409, conflict) which tells the client that it should back off for a
brief delay and try the same transaction again. Ideally that delay should
be randomized so that if several requests arrived roughly simultaneously,
the retries will be spread out over time a little bit and have a better
chance of succeeding. Look up "exponential backoff" on Wikipedia.

The random backoff does not need to be exponentially distributed, it can be
uniformly distributed and still work fine. The benefit of an exponential
distribution is that you're likelier to try sooner than later, so on average
you'll get the data a little sooner. According to the academic literature where
people have studied network collisions, exponential backoff is optimal. What I
did in some RueLaLa code was a linear-distributed backoff, like this.

 ```javascript
 var x = Math.random();     // uniform distribution between 0 and 1
 var y = 1 - Math.sqrt(x);     // linear distribution, decreasing
 setTimeout(function() { retry... }, y * 1000);   // max delay one second
 ```